MMV/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
sentence-transformers>=2.2.0
openai>=0.27.0
pillow>=8.3.0
opencv-python>=4.5.0
matplotlib>=3.4.0
numpy>=1.21.0
scipy>=1.7.0
scikit-image>=0.18.0
git+https://github.com/facebookresearch/segment-anything.git